merge to master

+++ src/vpn_service.py
@@ -18,25 +18,35 @@ if required. On each check the vpn_service sends some status information about t
 thermostats to the cloud, to keep the status information in the cloud in sync.
 """
 
+from __future__ import absolute_import
+
 from platform_utils import System
 System.import_libs()
 
+import glob
 import logging
 import os
-import glob
-import requests
-import time
 import subprocess
+import time
 import traceback
-import constants
-import ujson as json
-from threading import Thread, Lock
 from collections import deque
-from ConfigParser import ConfigParser
-from ioc import Injectable, INJECTED, Inject
-from gateway.config import ConfigurationController
+from threading import Thread
+
+import requests
+import six
+import ujson as json
+from six.moves.configparser import ConfigParser
+
+import constants
 from bus.om_bus_client import MessageClient
 from bus.om_bus_events import OMBusEvents
+from gateway.config import ConfigurationController
+from gateway.initialize import initialize
+from ioc import INJECTED, Inject
+
+
+if False:  # MYPY
+    from typing import Any, Deque, Dict, Optional
 
 REBOOT_TIMEOUT = 900
 DEFAULT_SLEEP_TIME = 30
@@ -141,7 +151,7 @@ class Cloud(object):
             if 'configuration' in data:
                 configuration_changed = cmp(self.__configuration, data['configuration']) != 0
                 if configuration_changed:
-                    for setting, value in data['configuration'].iteritems():
+                    for setting, value in data['configuration'].items():
                         self.__config.set(setting, value)
                     logger.info('configuration changed: {0}'.format(data['configuration']))
 
@@ -208,10 +218,10 @@ class Gateway(object):
             counters = data['counters']
 
             if self.__last_pulse_counters is None:
-                ret = [0 for _ in xrange(0, 24)]
+                ret = [0 for _ in range(0, 24)]
             else:
                 ret = [Gateway.__counter_diff(counters[i], self.__last_pulse_counters[i])
-                       for i in xrange(0, 24)]
+                       for i in range(0, 24)]
 
             self.__last_pulse_counters = counters
             return ret
@@ -245,7 +255,7 @@ class Gateway(object):
         """ Get the shutters status. """
         data = self.do_call("get_shutter_status?token=None")
         if data is not None and data['success']:
-            return [(int(shutter_id), details["state"].upper()) for shutter_id, details in data['detail'].iteritems()]
+            return [(int(shutter_id), details["state"].upper()) for shutter_id, details in six.iteritems(data['detail'])]
         return
 
     def get_thermostats(self):
@@ -322,6 +332,7 @@ class VPNService(object):
 
     @Inject
     def __init__(self, configuration_controller=INJECTED):
+        # type: (ConfigurationController) -> None
         config = ConfigParser()
         config.read(constants.get_config_file())
 
@@ -330,17 +341,17 @@ class VPNService(object):
         self._message_client.set_state_handler(self._check_state)
 
         self._iterations = 0
-        self._last_cycle = 0
+        self._last_cycle = 0.0
         self._cloud_enabled = True
-        self._sleep_time = 0
+        self._sleep_time = 0.0  # type: Optional[float]
         self._previous_sleep_time = 0
         self._vpn_open = False
         self._debug_data = {'energy': {},
-                            'master': {}}
-        self._eeprom_events = deque()
+                            'master': {}}  # type: Dict[str,Dict[float,Any]]
+        self._eeprom_events = deque()  # type: Deque[bool]
         self._gateway = Gateway()
         self._vpn_controller = VpnController()
-        self._config_controller = configuration_controller
+        self._config_controller = configuration_controller  # type: ConfigurationController
         self._cloud = Cloud(config.get('OpenMotics', 'vpn_check_url') % config.get('OpenMotics', 'uuid'),
                             self._message_client,
                             self._config_controller)
@@ -364,7 +375,7 @@ class VPNService(object):
         # If NTP date changes the time during a execution of a sub process this hangs forever.
         def popen_timeout(command, timeout):
             p = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-            for _ in xrange(timeout):
+            for _ in range(timeout):
                 time.sleep(1)
                 if p.poll() is not None:
                     stdout_data, stderr_data = p.communicate()
@@ -412,18 +423,18 @@ class VPNService(object):
         return False
 
     def _get_debug_dumps(self, dump_type):
-        if not self._config_controller.get('cloud_support', False):
-            return {}
+        # type: (str) -> Dict[float,Dict[str,Any]]
         debug_data = self._debug_data[dump_type]
         found_timestamps = []
         for filename in glob.glob('/tmp/debug_{0}_*.json'.format(dump_type)):
-            timestamp = int(filename.replace('/tmp/debug_{0}_'.format(dump_type), '').replace('.json', ''))
+            timestamp = os.path.getmtime(filename)
             if timestamp not in debug_data:
                 with open(filename, 'r') as debug_file:
                     try:
                         debug_data[timestamp] = json.load(debug_file)
                     except ValueError as ex:
                         logger.warning('Error parsing crash dump: {0}'.format(ex))
+                        debug_data[timestamp] = {}
             found_timestamps.append(timestamp)
         for timestamp in debug_data.keys():
             if timestamp not in found_timestamps:
@@ -431,12 +442,14 @@ class VPNService(object):
         return debug_data
 
     def _clean_debug_dumps(self, dump_type):
-        for timestamp in self._debug_data[dump_type]:
-            filename = '/tmp/debug_{0}_{1}.json'.format(dump_type, timestamp)
-            try:
-                os.remove(filename)
-            except Exception as ex:
-                logger.error('Could not remove debug file {0}: {1}'.format(filename, ex))
+        # type: (str) -> None
+        for filename in glob.glob('/tmp/debug_{0}_*.json'.format(dump_type)):
+            timestamp = os.path.getmtime(filename)
+            if timestamp in self._debug_data[dump_type]:
+                try:
+                    os.remove(filename)
+                except Exception as ex:
+                    logger.error('Could not remove debug file {0}: {1}'.format(filename, ex))
 
     @staticmethod
     def _get_gateway():
@@ -485,6 +498,7 @@ class VPNService(object):
         self._check_vpn()
 
     def _check_vpn(self):
+        # type: () -> None
         while True:
             self._last_cycle = time.time()
             try:
@@ -501,7 +515,7 @@ class VPNService(object):
                     time.sleep(DEFAULT_SLEEP_TIME)
                     continue
 
-                call_data = {'events': {}}
+                call_data = {'events': {}}  # type: Dict[str,Dict[str,Any]]
 
                 # Events  # TODO: Replace this by websocket events in the future
                 dirty_events = VPNService._unload_queue(self._eeprom_events)
@@ -518,7 +532,13 @@ class VPNService(object):
                 dumps = {}
                 dumps.update(self._get_debug_dumps('energy'))
                 dumps.update(self._get_debug_dumps('master'))
-                call_data['debug'] = {'dumps': dumps}
+
+                dump_info = {k: v.get('info', {}) for k, v in dumps.items()}
+                call_data['debug'] = {'dumps': {}, 'dump_info': dump_info}
+
+                # Include full dumps when support is enabled
+                if self._config_controller.get('cloud_support', False):
+                    call_data['debug']['dumps'] = dumps
 
                 # Send data to the cloud and see if the VPN should be opened
                 feedback = self._cloud.call_home(call_data)
@@ -552,10 +572,8 @@ class VPNService(object):
 
 if __name__ == '__main__':
     setup_logger()
+    initialize()
+
     logger.info("Starting VPN service")
-
-    Injectable.value(config_db=constants.get_config_database_file())
-    Injectable.value(config_db_lock=Lock())
-
     vpn_service = VPNService()
     vpn_service.start()
